Removed unnessecary command, fixed publisher, made copyps not fail

+++ psscripts/copyps.ps1
@@ -3,5 +3,9 @@ $CurrentLocation = Get-Location
 $DestinationDir = Join-Path $CurrentLocation $tsconfig.compilerOptions.outDir
 $SourceCodeDir = Join-Path $CurrentLocation $tsconfig.compilerOptions.rootDir
 $PowershellDir = Join-Path $SourceCodeDir "powershell"
+if (!(Test-Path $PowershellDir)) {
+  Write-Output "Nothing to copy."
+  exit
+}
 Copy-Item -Recurse -Force $PowershellDir $DestinationDir 
 Write-Output "Done copying from '$PowershellDir' to '$DestinationDir'"
